% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{close}
\alias{close}
\title{Create a Pane Closing Function}
\usage{
close(pkg_name)
}
\arguments{
\item{pkg_name}{the name of the package}
}
\value{
a function for closing the connection
}
\description{
Create a Pane Closing Function
}
\examples{
\dontrun{
close_connection <- pane_close("dplyr")
close_connection()
}
}
