% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{cook}
\alias{cook}
\title{Open the datasets from a package}
\usage{
cook(pkg_name)
}
\arguments{
\item{pkg_name}{the name of the package}
}
\value{
a function for closing the connection
}
\description{
Open the datasets from a package
}
\examples{
\dontrun{
cook("dplyr")
}
}
